+++ hardware/alchemy-lab/v2/src/alchemy_lab_v2_layout.cpp
@@ -39,20 +39,20 @@ struct QspiCtx { daisy::QSPIHandle* qspi; };
 
 QspiCtx s_qspi_ctx;
 
-bool v2_erase(void* ctx, uint32_t start, uint32_t end)
+bool v2_erase(void* ctx, uintptr_t start, uintptr_t end)
 {
     auto* c = static_cast<QspiCtx*>(ctx);
     return c->qspi->Erase(start, end) == daisy::QSPIHandle::Result::OK;
 }
 
-bool v2_write(void* ctx, uint32_t addr, const uint8_t* buf, uint32_t len)
+bool v2_write(void* ctx, uintptr_t addr, const uint8_t* buf, uint32_t len)
 {
     auto* c = static_cast<QspiCtx*>(ctx);
     return c->qspi->Write(addr, len, const_cast<uint8_t*>(buf))
            == daisy::QSPIHandle::Result::OK;
 }
 
-void v2_invalidate(void* /*ctx*/, uint32_t addr, uint32_t len)
+void v2_invalidate(void* /*ctx*/, uintptr_t addr, uint32_t len)
 {
     dsy_dma_invalidate_cache_for_buffer(
         reinterpret_cast<uint8_t*>(addr),
